refactor: move to be the encryption protocol
