--- src/encryption-state-vector/index.ts
@@ -1,1 +1,0 @@
-
