+++ src/lib/protocol/encryption/lamport-clock.ts
